refactor: reducer 함수 분리
